requirements.txt
torch==1.9.0
torchvision==0.10.0
h5py==3.5.0
pyyaml==6.0
matplotlib==3.4.3
scikit-learn==1.0.1
opencv-python==4.5.4.58
PyOpenGL==3.1.5
glfw==2.4.0
tqdm==4.62.3